6 size

+++ 0x00-hello_world/6-size.c
@@ -13,10 +13,10 @@ int main(void)
 	long long int z;
 	float f;
 
-	printf("Size of a char: %lu bytes(s)\n", sizeof(c));
-	printf("Size of an int: %lu bytes(s)\n", sizeof(i));
-	printf("Size of a long int: %lu bytes(s)\n", sizeof(l));
-	printf("Size of a long long int: %lu bytes(s)\n", sizeof(z));
-	printf("Size of a float: %lu bytes(s)\n", sizeof(f));
+	printf("Size of a char: %lu bytes(s)\n", (unsigned long)sizeof(c));
+	printf("Size of an int: %lu bytes(s)\n", (unsigned long)sizeof(i));
+	printf("Size of a long int: %lu bytes(s)\n", (unsigned long)sizeof(l));
+	printf("Size of a long long int: %lu bytes(s)\n", (unsigned long)sizeof(z));
+	printf("Size of a float: %lu bytes(s)\n", (unsigned long)sizeof(f));
 	return (0);
 }
